Improves Tests
I improved the tests to run several tests for each function, including testing edge cases. This will make sure the function are correct and reduce the probability of errors.

+++ src/main/python/testModule.py
@@ -30,7 +30,7 @@ def main(args):
 def add(args):
     sum = 0
     for num in args:
-        sum += int(num)
+        sum += float(num)
     return sum
 
 def optimiseReq(args):
